// src/lib/models/UserProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserProfile = {
    id?: number;
    nom?: string;
    prenom?: string;
    email?: string;
    role?: string;
    filiere?: string;
    annee?: number;
    statut?: string;
    interests?: Array<string>;
    connectionCount?: number;
    createdAt?: string;
};