// src/lib/models/EventResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EventResponse = {
    id?: number;
    titre?: string;
    description?: string;
    dateDebut?: string;
    dateFin?: string;
    lieu?: string;
    categoryId?: number;
    categoryNom?: string;
    categoryPriorite?: number;
    organisateurId?: number;
    organisateurNom?: string;
    status?: string;
    maxParticipants?: number;
    shareToken?: string;
    participantCount?: number;
    createdAt?: string;
    updatedAt?: string;
};